Refactor & use formatPkh everywhere
