+++ src/components/AccountSelector/AccountSmallTileDisplay.tsx
@@ -1,5 +1,5 @@
 import { Flex, FlexProps, Heading, Text } from "@chakra-ui/react";
-import { formatPkh } from "../../utils/formatPkh";
+import { formatPkh } from "../../utils/format";
 import { Identicon } from "../Identicon";
 import { prettyTezAmount } from "../../utils/format";
 import colors from "../../style/colors";
