+++ src/components/AddressTile/AddressTile.test.tsx
@@ -1,7 +1,7 @@
 import { mockMnemonicAccount } from "../../mocks/factories";
 import { render, screen } from "../../mocks/testUtils";
 import { Address } from "../../types/Address";
-import { formatPkh } from "../../utils/formatPkh";
+import { formatPkh } from "../../utils/format";
 import accountsSlice from "../../utils/redux/slices/accountsSlice";
 import { assetsActions } from "../../utils/redux/slices/assetsSlice";
 import store from "../../utils/redux/store";
