+++ src/components/OperationTile/TokenTransferTile.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from "../../mocks/testUtils";
 import { DefaultNetworks } from "../../types/Network";
 import { Token, fromRaw } from "../../types/Token";
 import { TokenTransfer } from "../../types/Transfer";
-import { formatPkh } from "../../utils/formatPkh";
+import { formatPkh } from "../../utils/format";
 import accountsSlice from "../../utils/redux/slices/accountsSlice";
 import { networksActions } from "../../utils/redux/slices/networks";
 import store from "../../utils/redux/store";
