--- src/utils/formatPkh.ts
@@ -1,3 +1,0 @@
-export const formatPkh = (pkh: string) => {
-  return `${pkh.slice(0, 5)}...${pkh.slice(-5, pkh.length)}`;
-};
